pipeline/pyproject.toml
[project]
name = "womens-mag-pipeline"
version = "0.1.0"
description = "Parses Russian entertainment sites and writes rewritten articles as markdown for the Astro site"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27",
    "feedparser>=6.0",
    "selectolax>=0.3.21",
    "pyyaml>=6.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "anthropic>=0.40",
    "python-dateutil>=2.9",
    "tenacity>=8.3",
    "eval-type-backport>=0.2",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.5"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"
